pyproject.toml
[project]
name = "articles"
version = "0.1.0"
description = ""
authors = [
    {name = "miloradik",email = "texetbuy@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "cryptography (>=46.0.7,<47.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
